Changes to Scenemanager and changed the distance to not be one byte

+++ Proftaak_Healthcare_B3/HealthcareClient/ServerConnection/DataManager.cs
@@ -51,7 +51,7 @@ namespace HealthcareClient.ServerConnection
             if (this.clientMessage.HasPage16)
                 PushMessage();
 
-            this.clientMessage.Distance = (byte)distance;
+            this.clientMessage.Distance = distance;
             this.clientMessage.Speed = (byte)speed;
             this.clientMessage.HasPage16 = true;
         }
